project/config/requirements/base.txt
asgiref==3.3.4
Django==2.2.4
djangorestframework==3.10.2
gunicorn==20.0.4
python-decouple==3.4
pytz==2018.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
